requirements_tyj.txt
python_dotenv==1.0.1
redis==5.0.6